refactor: Add database response saving to fanOutService.js
This commit adds a function to save the responses from downstream services to the database in the `fanOutService.js` file. The responses are now stored using the `DB` function, which inserts the data into the `responses` table. This update enhances the functionality of the application by persisting the responses for future reference.