refactor: Update fanOutService.js to save responses to the database
The fanOutService.js file has been updated to include a new function called saveToDatabase. This function is responsible for saving the responses from downstream services to the database. The responses are now stored using the saveToDatabase function, which inserts the data into the responses table. This update enhances the functionality of the application by persisting the responses for future reference.